Enum resolv::ResolverOption

pub enum ResolverOption {
    Init,
    Debug,
    AAOnly,
    UseVC,
    Primary,
    IgnTc,
    Recurse,
    DefNames,
    StayOpen,
    DNSrch,
    Insecure1,
    Insecure2,
    NoAliases,
    UseInet6,
    Rotate,
    NoCheckName,
    KeepTSIG,
    Blast,
    UseBString,
    NoIp6DotInt,
    UseEDNS0,
    SngLkup,
    SngLkupReop,
    UseDNSSEC,
    NoTLDQuery,
    Default,
}

Options for the Resolver

Variants
Init

address initialized

Debug

print debug messages

AAOnly

authoritative answers only (!IMPL)

UseVC

use virtual circuit

Primary

query primary server only (!IMPL)

IgnTc

ignore truncation errors

Recurse

recursion desired

DefNames

use default domain name

StayOpen

Keep TCP socket open

DNSrch

search up local domain tree

Insecure1

type 1 security disabled

Insecure2

type 2 security disabled

NoAliases

shuts off HOSTALIASES feature

UseInet6

use/map IPv6 in gethostbyname()

Rotate

rotate ns list after each query

NoCheckName

do not check names for sanity (!IMPL)

KeepTSIG

do not strip TSIG records

Blast

blast all recursive servers

UseBString

IPv6 reverse lookup with byte strings

NoIp6DotInt

Do not use .ip6.int in IPv6 reverse lookup

UseEDNS0

Use EDNS0.

SngLkup

one outstanding request at a time

SngLkupReop

one outstanding request at a time, but open new socket for each request

UseDNSSEC

use DNSSEC using OK bit in OPT

NoTLDQuery

Do not look up unqualified name as a TLD.

Default

Default values